refactor(core): rework texture update process
The main goal of this PR is to ease the usage of the texture update features of
itowns with objects that aren't TileMesh/LayeredMaterial.

One example of such situation is: being able to colorise a pointcloud using
a WMS server.

First the color and texture update mechanism have been splitted, since they don't
share that much code (yet).

Second, part of the update code has been moved back to the Provider and it helps
clarify the roles. Basically, the update loop has now no knowledge at all of
the internal logic of the behavior.

On the other hand the provider now directly implement the various update strategies,
since they're the only one capable of applying them to their protocol.

So the update process is now this;
   1. update code asks the provider if it wants to download new textures
   2. the provider returns the extents/url of the said textures
   3. the update code forward these commands to the scheduler
   4. when the command resolve, the update code process the result
(until now, the responsability of step 1/2 was shared between update/providers)

This also opens the way for a nice pattern: if at step 2 the provider returns the
full url the scheduler can directly check if the cache already has this resource
available and skip the whole queue/priority mechanism and improve the latency.