Add zAttenuation support to inpainting on points
Inpainting is a rendering technics that paints inside holes in the
pointcloud. One drawback of this: it also paints a small border to the
pointcloud. When we are far from the pointcloud, this border is really
disturbing, as its width does not depend on distance. This commit
gradually reduces the covered area according to distance.

By default, if the camera distance to the potential point is more than
100 meters, it is ignored in the inpainting process. Between 100 and
10 meters, it is progressively taken into account. Below 10 meters it is
fully taken into account.