3dtiles: fix min/max distance computation
This min / max is ultimately used in MainLoop to set the camera
near and far plane. Previously, we calculated a distance, then later we
added radius (sphere) or size (box) of node.boundingVolume.{sphere, box}.
BUT this dimension is affected by the node transformation matrix!! So we
need to calculate it when we have the transformed bounding volume.