apidoc: fixes hyperlinks to source files
Whenever the 'currentVersion' variable matches a semantic version,
the URL fragment to the git repository that matches the version will be
prefixed with 'v'. For example, if currentVersion = '0.2.5', the URL to
the source file Foo.js will be

    <repository>/tree/v0.2.5/src/Foo.js

However, if currentVersion does not match a semantic version, we use it
directly. For example, if currentVersion = 'master', the URL will be

    <repository>/tree/master/src/Foo.js