build(webpack): reorganize in dev and prod configs
Instead of a single webpack.config.js used in both production and
development, use a set of 3 webpack configurations :

- webpack.common.js
- webpack.dev.js
- webpack.prod.js

In package.json:

- `npm run start` uses webpack.dev.js
- `npm run build` uses webpack.prod.js.

webpack.common.js
=================

The common configuration for both development and production environments.

Specifies the entry and output, and the rules to apply.

Uses the babel-loader for transpiling.

webpack.dev.js
==============

This config does not perform any optimization and uses inline source maps.
The dev server is used to serve the examples.

webpack.prod.js
===============

This config perform tree shaking and minification and uses source maps.