build(examples): generate example page from template
The example page is automatically generated by webpack using a custom
plugin, example-builder.mjs.

Each HTML file in the examples/ directory is parsed to retrieve 2 <meta>
tags : name and description.

Then, for each example file, a HTML card is generated, using the name,
the description and the screenshot.

Finally, all HTML cards are aggregated and inserted in the index template
to generate the final page.