fix(DEMUtils.js): fix the logic of getting the correct tile we pick from
After the fix of #14, we can make the assumption that if one parent has received an elevation
texture, the current tile *will* have it, because the texture inheritance now happens every time
synchronously when creating children tiles.

So if we don't have an elevation texture, it means either there is none configured or no texture has
ever be received. In all cases, the picking will return a z of 0.