fix(LayeredMaterial.js): stop using arbitrary long timeout for refreshing atlas
For some reasons (?) we were waiting between 300 and 600ms to regenerate the atlas when new textures
were received. While keeping the setTimeout is useful to do the work in the next macro-task (which
avoids blocking the main thread too much), I don't see the point in slowing down stuff artificially.
Now textures loading are much more snappier (at least this is my perception).