fix(PotreePointCloud): computeScreenSpaceError() uses point radius instead of diameter
The screen space error computation uses the distance between points, minus their radius, to compute
the visible distance between points. So we must use the radius and not the "size", which is the
diameter.

BREAKING CHANGE: some adjustment may be needed in existing code to take into account the change in
the resulting SSE.