perf(OLTileProvider): produce DataTextures instead of Textures
DataTextures can be fed with the underlying pixel data in the canvas,
whereas regular Textures must be fed with an image element.

However, creating an image element from a canvas (with toDataURL()),
is very inefficient, because toDataURL() encodes the pixel buffer in PNG
and returns a string that will be fed in the image.src property.

feat(GeographicCanvas): add the getImageData() method

This method produces an ImageData, that can either be drawn into a canvas
with context.putImageData(), or put into a DataTexture.

It is much faster than getImage() since it does not perform any encoding or conversion.