perf(COGProvider): produce DataTextures instead of Textures
DataTextures can be fed with the underlying pixel data in the canvas,
whereas regular Textures must be fed with an image element.

However, creating an image element from a canvas (with toDataURL()),
is very inefficient, because toDataURL() encodes the pixel buffer in PNG
and returns a string that will be fed in the image.src property.