fix(Map): allow arbitrary EPSG:3857 extents
When a map CRS is EPSG:3857, we did not allow arbitrary extents,
but we adjusted the extent to fit into a tile grid. The rationale was
to ensure that a map tile mapped exactly to one source tile.

But this limitation is no longer useful because tile providers handle
arbitrary tile extents (by combining source tiles if necessary).

fix #42