feat(Map): subdivide extent into level 0 tiles to minimize distortion
Historically, the Map constructor supported an array of extents.

Those extents were supposed to be the root tiles (level 0 tiles).

However, this behaviour was not documented, and could lead to bugs if the
provided extents were not subdivisions of a rectangle (for example, two
extents that only shared a single corner).

Instead, the TileProvider automatically computes the best subvisions of
the extent to reduce the distortion of tiles.

For example, if a map has an extent that is twice as large as it is high,
it will be subdivided into two horizontal square level 0 tiles.

The benefits is that we have tiles that are more square shaped, which
helps visual quality (as generally we want tile texture to be square to
help with atlas fitting), as well as Screen Space Error computation,
as those algorithms work better on square or spherical objects.

BREAKING CHANGE : Map constructor no longer supports array of extents.