perf(OLTileProvider): further optimize the generation of tile textures
We can use a CanvasTexture to store the canvas to be drawn into the tile atlas, or use this texture
directly in the case of elevation layers. Now we lose the ability to recycle canvases, but we save
the whole conversion to ImageBitmaps and remove an asynchronous operation. The result is faster load
times for tiles, as well as a reduced memory usage for the provider.