fix(TileFS/HillShading): correct UV for hillshading
The origin UV for kernel computation of slope/hillshade was wrong. This fix correct it using an
already available method. To test the fix one can look at the hillshade example to see that it is
still correct and set "hillshading: true" to the map of customtiledimage example (l.27) to see that
issue #91 is resolved.

fix #91