fix(OLTileProvider): don't override layer.extent
OLTileProvider was overriding the layer extent in the preprocessing step, causing a layer extent to
be much much bigger than the map extent, causing issues in other preprocessing steps, such as
min/max computation, where the whole layer extent was downloaded (instead of the map extent).

Only set extent if it is not set at all.

fix #94