fix(LayeredMaterial.js): remove delayed repaint of the atlas
Whenever a new color layer texture is available for a tile, the LayeredMaterial
must repaint the atlas with the new texture. Normally, only the new texture
is redrawn, but it actually never worked, because we forced all textures
to be redrawn again. A warning was logged every time a repaint occurred,
but this warning made no sense.

This fix removes :
- the delayed redraw (used a setTimeout with 1 ms delay)
- the warning
- the notion of "pending updates" that was obsolete because all layers
  were always repainted anyway

fix #99