feat(TextureGenerator): replace nodata values by valid values to remove interpolation effect
Update nodata with first or last known valid value. This removes the unwanted interpolation at the
interface between nodata and valid values.

We check if a nodata value has neighbors with elevation. If this is the case we change the elevation
to be the neighbor one. Even if the nodata pixel is discarded because of opacity=0.0, the
interpolation that remained is now between pixels of similar elevation, removing the "cliff" effect.

re #119