fix(LayeredMaterial): handle direct-to-canvas drawing before clearing the atlas
Some providers (OLVectorTileProvider for example) don't produce images
to be drawn into the atlas. Rather, they draw the vectors directly in the
atlas.

They would produce a null image, indicating that the data was directly
drawn into the atlas, and that drawLayerOnCanvas should do nothing.

However, clearing the atlas would check if the image was != undefined,
leading to a non wanted clear, since the image was null and not undefined.

fix #124