feat(instance): improve `instance.focusObject` and add it to the API
This method will try its best to put the current object entirely into
view. Due to some approximation, it should be *mostly* the case.

It improves from the previous version by supporting all kind of object.
The previous version was somewhat inconsistent because for Map it
actually implemented a topdown view.