feat: support all pixel types for composed textures
Introduce the WebGLComposer class, that can compose several image tiles
into a single texture. Contrary to the previous implementation, this
does not use a canvas, but a WebGL context to draw rectangular meshes
into an orthographic camera's frustum, then rendering the scene into
a texture.

This enables support for arbitrary pixel types in composed textures,
whereas we were previously limited to 8-bit RGB images.

The previous implementation (which was in GeographicCanvas) still exists,
but has been renamed CanvasComposer. The APIs for both implementations are
identical.

There is also the Composer class, that is an abstraction over both
implementations. This class automatically chooses the best implementation
depending on the textures/images that are drawn on the composer area.

Two classes are affected by the changes :
- LayeredMaterial, which now uses a WebGLComposer to render its atlas
- OLTileProvider, which uses the Composer to compose its tiles