fix(OLTileProvider): return DATA_UNAVAILABLE when below the min zoom level
For layers that are defined with partial zoom levels (i.e not starting at 0),
the naive implementation would use the first available zoom level of the
source to fill the map tile.

However, this could lead to potentially thousands or millions of tile
requests when the difference between the theoretical and actual zoom
level is big.

We now determine that if the minimum zoom level of the source is still
too high for our needs, we return DATA_UNAVAILABLE.

Conversely, instead of returning null when the theoretical zoom level is
greater than the maximum available zoom level, we return the maximum
available zoom level, because this is harmless.

fix #146