refactor(TMSProvider): remove obsolete TMSProvider to use ol sources
From some time, we have been using ol sources for every standard raster
protocol. Ol provides sources for TMS and XYZ protocol, making the
existence of TMSProvider not needed any more.

Note: this is not a breaking change, because it was not possible to use
TMSProvider for a long time already (at least since 5ab0ba8e3).