feat: support xbil elevation format from IGN
We used to support it, but for lack of example, this has been broken.
This commit brings back the support, following the new structure of the
code and especially integrating it with TextureGenerator.

The parsing is done by a custom format, but after that it is treated as
a NUMERIC ElevationLayer.