fix(Map): support stitching including the 4 diagonal neighbours
Stitching used to only support the 4 edge neighbours (T, R, B, L), leading to some
visible cracks on corners.

This commit extends the stitching logic to the 4 additional corner (diagonal)
neighbours (TL, TR, BL, BR):

   TL  T  TR
    L  *  R
   BL  B  BR

This also refactors the way the map processes the neighbours, by moving
this code into the TileMesh.

The names of the stitching-related uniforms are updated to be more explicit.

Finally, this fixes a bug in the TileIndex, where a tile could have itself
for neighbour.

fix #134