refactor: decode elevation formats in the provider
Removal of ELEVATION_FORMAT
===========================

The ELEVATION_FORMAT enum is removed to be replaced by the Interpretation
class, that describes how the pixels of an image are to be interpreted :

    - Interpretation.Raw : the pixels are used without transformation
    - Interpretation.MapboxTerrainRGB: the pixels encode an elevation in
      a color image using the Mapbox Terrain RGB scheme
    - Interpretation.ScaleToMinMax(min, max): scale the pixels using
      the provided min and max values.

This interpretation is now a property of all layers, and not exclusively
elevation layers.

Interpretation of the data
==========================

Previously, interpreting the values in elevation textures was done in
the vertex and fragment shaders of the map tiles.

Now, this is done upstream, during the generation of the texture in the
composer itself.

This simplifies greatly the tile shaders, and avoid recomputing the
interpretation in each vertex/fragment.

WebGLComposer
=============

The WebGLComposer now include logic to detect the best data type for the
generated textures, based on the interpretation.

For example, if the input texture is a 8-bit color image with
Interpretation.MapboxTerrainRGB, it means that the output texture must be
in the FloatType data type to be able to contain the elevation values.

BREAKING CHANGE: The property `elevationFormat` of the `ElevationLayer`
is replaced by the `interpretation` option in the `Layer` class:

    ```js
    // instead of
    const layer = new ElevationLayer(..., { elevationFormat: ELEVATION_FORMAT.NUMERIC });

    // do
    const layer = new ElevationLayer(..., { interpretation: Interpretation.Raw });
    ```