fix(TileVS): fix corner stitching when neighbours are bigger than the current tile
There was a missing branch in the corner stitching feature : the case
when the neighbours are bigger than the current tile.

In this case, we cannot average the elevation of the current tile with
the neighbours, but we have to only take the neighbours' elevation. The
reason is because the bigger neighbour will not average with our elevation,
creating a difference in elevation, and an incorrect stitching.

Also, in the case that the neighbours have different sizes (but still
bigger than our tile), we only consider the biggest neighbours.

For example, if our tile is level 5, and it shares a corner with 3
neighbours, of level 5, 4, and 3, we only take the elevation of neighbour
with level 3.

fix #164