build(npm): add a start script
My frontend dev muscle memory always want to type `npm start` and then I remember that we don't have
this in giro3d. This is a very common name for "starting the default stuff" on js project (so common
that you don't even need to use `run` in the command) so I propose we alias it to serve-examples.