fix(Map.js): don't artificially stop division when tile size is < 5
Instead, consider a default maxSubdivisionLevel of 30, which should
cover all use cases.

Note that for 3857, things starts to go hairy before this level, because
the extents becomes too small for some calculation to work properly (for
instance those involving dimensions), but at least it does not fail or
worse, enter an infinite loop (which happens when you remove all
bounds).