feat: add the HttpConfiguration module
This commit introduces the HttpConfiguration module.

It is now possible to specify headers to apply to HTTP requests on a
per-host, and per-URL prefix basis. The longest prefix will have
precedence and will be evaluated first. Whenever a header has already
been set by a higher precedence prefix, or directly by the user, a lower
precedence prefix will not set the same header.

Examples
========

If the user wishes to use a specific header for a specific domain for
example:

    HttpConfiguration.setHeader(
        "example.com",
        "Authorization",
        "Bearer XXXX",
    );

If the user wants to use a different credential for a sub resource in
the same domain:

    HttpConfiguration.setHeader(
        "example.com/sub/resource",
        "Authorization",
        "Bearer YYYY",
    );

The longest prefix will have precedence and all URLs that match
"example.com/sub/resource" will use the "Bearer YYYY" Authorization.

fix #86