fix(Map): set a limit on the aspect ratio
Currently, when computing the root tiles of the Map, we use the aspect
ratio of the extent (dims.x / dims.y) to determine how many root tiles
to put on each axis.

Root tiles should be as square as possible to improve performance and
data quality.

If the ratio is really big, we are going to create a *lot* of root tiles
in an axis.

This patch sets a limit of a ratio of 10 before we stop trying to create
square root tiles, and we revert to rectangular root tiles.

fix #144