fix(OLTileProvider): use tile ratio for zoom level
When computing the appropriate zoom level in an OpenLayer tile grid,
we previously used the width of the extent, ignoring the height.

This was because we assumed that the extent was mostly square, so it
didn't make a difference. However, now that Map tiles stop being square
if the Map extent is really elongated, we must take into account this
extent ratio to compute the appropriate zoom level, otherwise the
computed zoom level would be way too big, and lead to memory exhaustion.

Now we use the biggest side of the tile to compute the zoom level.

fix #144