fix(WebGLComposer): don't clear() the renderer
The renderer does not belong to the WebGLComposer, it is actually the
renderer for the whole instance of Giro3D. Clearing it means modifying
the framebuffer that will render the whole canvas.

WebGLComposer only borrows this renderer to perform renders.

This is probably a relic from the previous implementation of
WebGLComposer that used its own WebGLRenderer.