fix(COGProvider): stop compressing data to 8-bit
If the COG was applied to a Color layer, the pixel data would be
compressed to 8-bit so that we could properly display non 8-bit data,
for example elevation data, into a color layer.

However, this behaviour prevented the use of color maps, since the
min/max values of the layer would not match with the compressed pixel
values.

Now, we stop forcing this behaviour. To display non 8-bit data as color
information, the user can use a color map (with a grayscale gradient).

fix #216