perf(Map): don't keep data copy of elevation textures
For all elevation textures, we previously stored a copy of the texture
in CPU memory (because the texture itself only exists in GPU memory).

This was done for 2 reasons :

    1. to sample the elevation texture during picking
    2. computing the min/max of the texture to get the tile bounding box

(1) is no longer relevant, as GPU picking samples the elevation texture
directly in GPU, (2) still remains.

Min/Max computation
===================

Now, min/max computation is done using either one method :

    1. compute min/max on source data (e.g BIL tiles, TIFF tiles, COG buffers)

    2. if the previous method didn't work (because it was not implemented
       in the format, or for any other reason), fallback to min/max
       computation on the texture rendered by the WebGLComposer,
       using a temporary copy.

The best method is (1), because generating a copy from the texture
generated by the WebGLComposer is costly (it requires a readback from
GPU memory).

Of course this copy is immediately deleted after min/max computation,
so we avoid the memory footprint, but the latency cost in the update
loop is still high.

Benefits
========

- The memory footpring of a scene with one or multiple Maps is greatly
reduced.
- If min/max computation is done on source data, we don't even create
the temporary copy, saving lots of precious milliseconds in update loop.

fix #215