perf(Cache): prevent unbounded growth
This commit rewrites the Cache to ensure that it cannot grow indefinately.

The Cache is now an instantiable class. To maintain compatibility with
the previous singleton-based approach, a GlobalCache object is exposed,
that acts as the only instance of the Cache class. This singleton can
be removed in the future if we choose to instantiate one Cache per
Instance.

The underlying implementation uses the `lru-cache` package, that provides
features to set limits to the cache growth :
- max number of entries
- max total size of entries (in arbitrary units, but we use bytes)
- TTL (time to live) of entries

The Cache class provides default values for those 3 limits, but they
can be changed.

In effect: the Cache class is just a thin wrapper around the LRUCache
class, that enables us to easily switch the underlying implementation,
or enforce various policies that LRUCache does not care about.

One of those policies is that the Cache class expects cache keys to be
strings, whereas LRUCache does not care. Having strings as keys is useful
for various reasons: it's easy to ensure the uniqueness of the keys, we
avoid potential memory leak by using disposable objects as keys, and
the keys can be self-descriptive, which helps during debugging.

The API is mostly the same, with the following changes:
- the flush() method is removed, as LRUCache does it automatically
- the set() method takes an optional parameter for all the options

fix #225