perf(COGProvider): handle command cancellation
Add cancellation points during the COG processing pipeline :
- before starting the processing
- after fetching the data blocks and before processing the images

Before actually processing the command, we wait 250ms to ensure that
the command is really needed, as COG processing is resource heavy.

This can save a huge amount of resources because a lot of commands
get actually cancelled (when their source becomes hidden for example,
which happens a lot during fast camera movements).

fix #234