fix(TextureGenerator): fix forever pending create8bitImage()
create8bitImage(), now called load8bitImage(), previously created its
own HTMLImageElement before loading it with the provided blob. On some
browsers, the image would sometime never load, indicating a possible
elimination of the image by the browser's garbage collector.

Now, we pass the image element to the function so that it is not
eliminated.

fix #232