feat(Provider): providers can decide extents and texture size
Providers are now free to:
- use a different extent than the requested extent
- use a different texture size than layer.imageSize

Of course, within reason.

This allows "pixel perfect" textures, i.e textures whose pixels are
aligned with the source pixels (without redimensioning pixels nor cutting
them).

For example, if the requested extent cuts in the middle of source pixels,
then we adjust this extent so that its edges are also on the edge of
pixels.

This mitigates a number of issues when zooming in beyond the maximum
resolution of the source: aliasing issues, "jumping" pixels, etc.

fix #243