feat: add elevationRange on Map and ColorLayer
This optional feature can be used to limit the display of a ColorLayer
within an elevation range. The ColorLayer will not contribute to the
tile color when the pixel is below or above this range.

If the feature is enabled on the Map, then the whole map is displayed
only within the range.

The two features can be combined.

fix #246