fix(Map): don't assign imageSize to layers
Layers are mostly shareable between maps, with the exception of the
imageSize property that is set by the Map when it is attached.

Remove this property and put it in the map tiles/point cloud tiles
instead, so that layers are no longer altered by the map.

Refactor the providers to honor this new parameter in getPossibleTextureImprovements().

fix #257