fix(Picking): render with a black clear color to avoid decoding issues
GPU picking renders a tiny fraction of the viewport around the picked
coordinate, in order to determine what object was picked. The object is
determined by decoding the rendered color.

However, we need to ensure that the clear color (background color) of
the rendered viewport is black to avoid interfering with color decoding.

fix #249