fix(composer): fix no-data filling algorithm
The previous algorithm (executed in TextureGenerator), had a flaw in
some cases, where no-data pixels would not properly sample the nearest
valid pixel, and instead sample a much farther pixel, resulting in
steep walls at the edge of the data.

A new algorithm is used instead, and for performance improvements, is
executed in the composer shader (ComposerTileFS.glsl), using the
FillNoData.glsl chunk.

The new algorithm works by sampling the valid pixel that is the closest
to the position. This is quite similar to the FillNoData() function of
GDAL (https://gdal.org/programs/gdal_fillnodata.html), the difference
being that we cannot affort to sample all pixels to maintain correct
performance, so we use a fixed number of samples that is less than
the number of pixels in the image. However the result is quite correct.

To further save CPU time, the ComposerTileMaterial is recycled as much
as possible between invocations, to avoid costly calls to WebGL, which
take around 80ms per tile !

fix #151