perf(Map): process picking in a single pass
GPU picking for maps require 3 separate informations:
- the tile UUID
- the UV
- the elevation of the picked point

Previously, this was done in 3 separate render passes.

Now, the 3 infos are rendered in a single pass, using a 32-bit render
target. As a bonus, this avoids encoding and decoding the data.

As a result, this improves performance of picking map tiles, by
avoiding GPU stalling due to numerous readPixels() invocations.

fix #274