perf(LayeredMaterial): repaint the atlas at most once per frame
Currently, the atlas is repainted every time a color texture is set
on the material. Thus, if the material has N color layers, the atlas
is repainted N times per frame.

To save precious GPU time, the atlas is now repainted at most once,
and only if the mesh is actually rendered, using the onBeforeRender()
API of three.js.