refactor(Layer): rewrite layers
This commit rewrites the mechanism of layers (both ColorLayers and ElevationLayers)
with a simpler and more flexible approach.

Summary
=======

1. Merge providers and sources

Providers are removed and their code is merged with the sources.

For sources that come directly from OpenLayers, three wrappers are
created:
- OLTiledImageSource for tiled images
- OLVectorSource for all kinds of vectors
- OLVectorTileSource for vector tiles

Layer sources must implement the ImageSource base class.

2. Introduce the LayerComposer

The main change resides in the introduction of the LayerComposer class,
in the layer itself. This class is responsible for aggregating and composing
the images coming from the sources. Indeed, sources are no longer responsible
for producing a single image with the correct extent (they were using
the WebGLComposer for that). Now, sources can output all the images that
they want for the requested extent, and the LayerComposer will compose them.

3. Reduce overall coupling

Sources are no longer coupled to layers. Layers are weakly coupled to
sources (only via the base class ImageSource).

Layers are very weakly coupled to their parent entity.

4. Performance optimizations

A major performance bottleneck was the computation of min/max values
for elevation textures, that was done in the WebGLComposer (and
required a costly texture readback).

Now min/max are computed on the source images (produced by the sources),
which is much cheaper.

BREAKING CHANGE: The API for sources and layers has changed. See the
documentation for more information.