perf: avoid creating DataTextures when possible
Creating DataTextures from remote images requires a readback of the
image data into a buffer. It may be necessary if we need the pixel buffer,
i.e to compute the min/max elevation.

However, we generally don't need this pixel buffer, and we can directly
create the texture from the image.

fix #336