feat(Map): allows limiting hillshading to elevation layers only
The `hillshading` parameter of the Map constructor is now either a
boolean or an object of type `HillshadingOptions` that contains all the
parameters related to shading.

In the case of a boolean, `true` enables hillshading with default values,
and `false` disables hillshading:

Boolean parameter:

```js
const map = new Map('boolean', { hillshading: true });
```

Object parameter:

```js
const map = new Map('object', {
    hillshading: {
        enabled: true,             // enables hillshading on this map
        elevationLayersOnly: true, // avoid shading color layers
        zenith: 90                 // sun at zenith (90 degrees)
        azimuth: 0                 // sun rays are coming from the north (0 degrees)
    },
});
```

BREAKING CHANGE: `Map.materialOptions.hillshading` is now an object of type
`HillshadingOptions` rather than a boolean. `Map.lightDirection` is removed
and the `zenith` and `azimuth` parameters are now part of the `HillshadingOptions`
object.