feat(CogSource): support channel mapping
A COG file is a regular GeoTIFF that can contain many bands (or samples).
However, in the previous implementation, the user had no control over
which band to read, and the implementation would read *all* bands in
order, and interpret them as [R, G, B, A] (in the case of 4 bands or
more).

This would lead to inefficient reading of unused bands, as well as
incorrect interpretation of the 4th band as an alpha channel, without
any certainty that this channel did actually contain alpha information.

This would also prevent the user to remap bands as it is frequent in
GIS packages, for example to display an infrared band as the green
channel.

Now, the CogSource exposes a `channels` option in the constructor where
the user can specify how to map source bands/samples to target channels.

fix #335