fix(CogSource): select correct data type from sample format and bits per sample
To select the best container for COG pixels (unsigned byte, float, etc.)
we were using an incorrect method: the number of bytes per pixel (BPP).
But this method gives the same result for a single band float (1 * 4 BPP)
and a 4 band 8-bit image (4 * 1 BPP as well), so they would both map to
UnsignedByteType, and we would lose information.

The new method is taken directly from geotiff.js source code, and maps
8-bit pixels to UnsignedByteType, and everything else to FloatType.

fix #333