fix(ImageSource): createReadableTextures is now a parameter of getImages()
In the (rare) case where the same source is shared between and elevation
layer and a color layer, the parameter `createReadableTextures` that
was set in the `initialize()` method would contradict the needs of both
layers: the color layer would set it to false, and the elevation layer
would set it to true. But since the source can be initialized only once,
the first layer to initialize it would "win".

Now, this parameter is passed for every call to `getImage()` to avoid
this conflict.