fix: correctly set .needsUpdate when changing opacity on Entities
material.needsUpdate is not a real properties, it's only a setter. As a
result, reading `material.needsUpdate` always gives undefined. So we
can't use pattern like:

```js
material.needsUpdate |= foo;
```

Instead we should use `||=` or simply an assignment. I chose the latter
here because I don't like the first solution with a LHS that is always
undefined.