fix(Layer): raise an error if the extent's CRS does not match the instance's
The Layer does not know about the instance CRS up until preprocess(),
meaning that even though it could convert the extent there, in the time
between the layer creation and the preprocess() call, the extent might
be invalid, and the user could be using it. This could lead to all
kind of issues down the line.

To be safe, we simply reject an extent that does not match the instance CRS
and encourage the user to convert it beforehand.