fix(Layer): fix texture inheritance (#359)
When looking for a potential ancestor to inherit a texture from, we loop
over all tiles and select one whose extent contains the target extent.

However, the tested extents are not precise subdivisions in a quadtree
fashion: the extents are actually altered to better fit the source's
pixels. This causes some tiles, which are logically descendant of a tile,
to not have an extent that "is inside" any of the logical ancestors.

This happens especially at the edges of the layer.

This is fixed by comparing the unaltered, geometrically precise extent
of tile.

fix #359