feat(Layer): enable changing the resolution of the layer (#301)
Previously, the size of the textures applied to map tiles were decided
by the map itself, based on a base size of 256*256 pixels (adjusted
to fit the aspect ratio of the tiles).

Now, it is also possible to oversample or undersample individual layers
in order to have a better resolution.

fix #301