fix(LayerComposer): apply no-data filling as post-process
Previously, the no-data filling was part of the preprocessing of images
that were added to the composer.

However, due to some images being reprojected, the no-data filling would
not take into account new empty areas created by the reprojection.

By moving the no-data filling as a post-processing (after reprojection),
the artifacts disappear.