refactor(Layer): remove id constructor parameter and introduce name
Previously, the constructor of `Layer` would take a mandatory id
parameter, that was supposed to be unique. However, there was no way
to guarantee that. Besides, this identifier is never used anywhere
in the library, except for checking that a layer was already added to a
map.

Now, this parameter is removed, and the `id` property of the Layer is
an automatically generated UUID.

The constructor now also takes an optional `name` parameter, which is used
as a user-friendly identifier, and does not have to be unique.

BREAKING CHANGE: The `id` parameter of Layer constructor is removed.