perf(Map): use 2 channels instead of 4 for elevation textures (#376)
The whole processing chain from `ImageSource`s to the `LayeredMaterial`
now uses textures that are as small as possible to accomodate for the
desired data type.

For elevation textures, we now use 2 channels instead of 4, in the
`RGFormat` pixel type, rather than `RGBAFormat`.

This means that we save 50% of memory usage per elevation texture,
but we have to be aware that the transparency value is now in the green
channel rather than the alpha channel. The shader code that process
elevation textures is updated to reflect that.

For color layers, in the case of grayscale images, we perform a RGB
expansion from grayscale in the `ComposerTileFS` shader, because the
original texture being in RGFormat, it will appear red instead of gray
without this expansion.

fix #376