fix(Map): guarantee that addLayer() always resolve (#387)
In some (rare) cases involving multiple concurrent calls to `addLayer()`
on different `Map`s for the same layer, some calls to `addLayer()` might
not resolve and the promise would be forever in the pending state. This
was due to a faulty mechanism to prevent adding the same layer twice.

Now this mechanism is removed.

This commit also uses this opportunity to remove the `attachedLayers`
array from the `Entity3D`, and let entities handle their layers, if any.

To help with this, a new interface, `HasLayers`, is introduced. It is
implemented by `Map`, `Tiles3D` and `PotreePointCloud` and provide
high-level methods to query layers in the object.

It is used by `MainLoop` to iterate and update layers.

BREAKING CHANGE: `Layer.whenReady` is removed. If you want to await
the initialization of the layer, just do `await Layer.initialize()`. Note
that this should not be necessary, as the layer is initialized when
added to the map.

fix #387