refactor(Instance): remove `layers-initialized`
This event was never triggered. Moreover, there is no reliable way to
get the exact moment when entities (the "layers" that the event refers to)
can be initialized, as entities can be added at any moment to the Instance.

BREAKING CHANGE: The event `layer-initialized` is removed from the `Instance`,
as it was never triggered.