refactor(c3DEngine): remove ColorManagement option (#373)
This option was added as an escape hatch to disable automatic color
management in THREE.js, but this option is now incompatible with the
recent changes in color handling, and would create issues.

BREAKING CHANGE: The `colorManagement` option in the `Instance` constructor
is removed. THREE.js color management is now always enabled.