perf(ImageFormat): optionally return min/max of texture
Some `ImageSource` can compute the min/max of a texture with no
additional cost (e.g BilFormat already loops over the pixel buffer to
process no-data values), so that we can avoid computing it later in the
processing chain.

BREAKING CHANGE: the return type of `ImageSource.decode()` has changed.
It initially returned a `Texture`, but it now returns an object of the
following type:

```ts
{
    texture: Texture;
    min?: number;
    max?: number;
}
```