feat(CogSource): support other colorspaces than RGB (#416)
Using the geotiff.js readRGB() call, we can automatically convert to RGB
pixel data that is in a different color space, such as paletted, YMCK,
or YCbCr. The `CogSource` exposes a new constructor option, `convertToRGB`,
that internally will use the `readRGB()` call rather than `readRaster()`
to perform the conversion.

fix #416