perf(LayeredMaterial): avoid using the atlas if possible (#417)
Previously, all color textures (from color layers) would be aggregated
into an atlas. This was done to circumvent the maximum number of textures
allowed by WebGL, which depends on the actual hardware.

This number of texture units can be accessed with

    gl.getParameter(gl.MAX_TEXTURE_IMAGE_UNITS)

Now, we only use the atlas texture if the number of texture units is
insufficient to accomodate all the textures used by the material.

Otherwise, we use each individual color texture directly in the shader.

This results in much greater performance (because we avoid the time
spent rendering the atlas), and memory usage(because we don't have to
allocate memory for the atlas texture).

fix #417