perf(LayerComposer): reduce the use of the pre-processing stage
The pre-processing stage was originally used to perform no-data filling,
texture flipping, elevation interpretation decoding and RGB expansion
of grayscale images.

However, the no-data filling algorithm has been moved to post-processing,
and the remaining operations are so cheap that they can be done in the
main stage (the render stage).

This removes the allocation of a `RenderTarget` for each image of the
composer and reduce the number of GPU operations required.

The only exception is for elevation textures that require a change in
filtering mode (nearest neighbour to linear).